Updated Apple Silicon Support and support for blender 3.6
This update adds a conditional that detects if the system is using Apple Silicon.

If the system is not using Apple Silicon, it will just default to the regular previous code, if it is, it is using the new code added by nurjeff.

The plugin worked for Blender 3.6 by default, but now I have updated the _init_.py to reflect it.

Co-Authored-By: Jeff Schmitz <57395948+nurjeff@users.noreply.github.com>